requirements.txt
streamlit>=1.7.0
requests>=2.26.0
flask>=2.0.1
Pillow==8.4.0